Move TransactionBuilder to testing package
