+++ core/trino-main/src/test/java/io/trino/security/TestAccessControlManager.java
@@ -67,8 +67,8 @@ import static io.trino.SessionTestUtils.TEST_SESSION;
 import static io.trino.spi.security.AccessDeniedException.denySelectTable;
 import static io.trino.testing.TestingEventListenerManager.emptyEventListenerManager;
 import static io.trino.testing.TestingHandles.TEST_CATALOG_NAME;
+import static io.trino.testing.TransactionBuilder.transaction;
 import static io.trino.transaction.InMemoryTransactionManager.createTestTransactionManager;
-import static io.trino.transaction.TransactionBuilder.transaction;
 import static java.nio.file.Files.createTempFile;
 import static java.nio.file.StandardOpenOption.CREATE;
 import static java.nio.file.StandardOpenOption.TRUNCATE_EXISTING;
