+++ plugin/trino-kinesis/src/test/java/io/trino/plugin/kinesis/TestRecordAccess.java
@@ -43,7 +43,7 @@ import java.util.UUID;
 import java.util.zip.GZIPOutputStream;
 
 import static io.trino.testing.TestingSession.testSessionBuilder;
-import static io.trino.transaction.TransactionBuilder.transaction;
+import static io.trino.testing.TransactionBuilder.transaction;
 import static java.lang.String.format;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.assertj.core.api.Assertions.assertThat;
