+++ testing/trino-testing/src/main/java/io/trino/testing/statistics/MetricComparator.java
@@ -28,7 +28,7 @@ import java.util.OptionalDouble;
 
 import static com.google.common.collect.ImmutableList.toImmutableList;
 import static com.google.common.collect.Iterables.getOnlyElement;
-import static io.trino.transaction.TransactionBuilder.transaction;
+import static io.trino.testing.TransactionBuilder.transaction;
 import static java.lang.String.format;
 import static java.util.stream.Collectors.joining;
