+++ testing/trino-tests/src/test/java/io/trino/tests/TestMetadataManager.java
@@ -30,9 +30,9 @@ import io.trino.spi.connector.ConnectorViewDefinition;
 import io.trino.spi.connector.SchemaTableName;
 import io.trino.testing.DistributedQueryRunner;
 import io.trino.testing.TestingSessionContext;
+import io.trino.testing.TransactionBuilder;
 import io.trino.tests.tpch.TpchQueryRunnerBuilder;
 import io.trino.tracing.TracingMetadata;
-import io.trino.transaction.TransactionBuilder;
 import org.intellij.lang.annotations.Language;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
